updated cancel pending request
